Tonton Maxence

+++ src/ml/struct/layer.c
@@ -2,7 +2,7 @@
 #include "layer.h"
 
 
-struct Layer CreateLayer(struct Neurone neurones[], int nb_neurones) {
+struct Layer CreateLayer(struct Neurone neurones[], size_t nb_neurones) {
     struct Layer layer;
 
     layer.nb_neurones = nb_neurones;
